Updates from feedback from Chris
