+++ api/src/main/java/ca/bc/gov/educ/api/gradstudent/model/dto/StudentCoursePagination.java
@@ -1,11 +1,10 @@
 package ca.bc.gov.educ.api.gradstudent.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import jakarta.validation.constraints.*;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
 
 import java.util.UUID;
 
@@ -20,15 +19,9 @@ public class StudentCoursePagination extends BaseModel {
     private String courseID;
     @NotBlank
     private String courseSession;
-    private Integer interimPercent;
-    private String interimLetterGrade;
     private Integer finalPercent;
-    private String finalLetterGrade;
     private Integer credits;
     private String equivOrChallenge;
-    private String fineArtsAppliedSkills;
-    private String customizedCourseName;
     private UUID studentExamId;
-    private String relatedCourseId;
 
 }
