+++ api/src/main/java/ca/bc/gov/educ/api/gradstudent/model/entity/StudentCoursePaginationEntity.java
@@ -34,17 +34,8 @@ public class StudentCoursePaginationEntity extends BaseEntity {
     @Column(name = "COURSE_SESSION", nullable = false)
     private String courseSession;
 
-    @Column(name = "INTERIM_PERCENT")
-    private Double interimPercent;
-
-    @Column(name = "INTERIM_LETTER_GRADE")
-    private String interimLetterGrade;
-
     @Column(name = "FINAL_PERCENT")
     private Double completedCoursePercentage;
-
-    @Column(name = "FINAL_LETTER_GRADE")
-    private String completedCourseLetterGrade;
 
     @Column(name = "NUMBER_CREDITS")
     private Integer credits;
@@ -52,16 +43,7 @@ public class StudentCoursePaginationEntity extends BaseEntity {
     @Column(name = "EQUIVALENT_OR_CHALLENGE_CODE")
     private String equivOrChallenge;
 
-    @Column(name = "FINE_ARTS_APPLIED_SKILLS_CODE")
-    private String fineArtsAppliedSkills;
-
-    @Column(name = "CUSTOM_COURSE_NAME")
-    private String customizedCourseName;
-
     @Column(name = "STUDENT_COURSE_EXAM_ID")
     private UUID studentExamId;
-
-    @Column(name = "RELATED_COURSE_ID")
-    private BigInteger relatedCourseId;
 
 }
